+++ src/components/base.ts
@@ -1,0 +1,26 @@
+
+//Component class
+export abstract class Component<T extends HTMLElement,U extends HTMLElement> {
+  templateElement: HTMLTemplateElement;
+  hostElement: T;
+  element: U;
+
+  constructor(templateId: string, hostElementId: string,insertAtStart: boolean, newElementId?: string) {
+    this.templateElement = document.getElementById(templateId)! as HTMLTemplateElement;
+    this.hostElement = document.getElementById(hostElementId)! as T;
+
+    const importedNode = document.importNode(this.templateElement.content, true)
+    this.element = importedNode.firstElementChild as U;
+    if (newElementId) {
+      this.element.id = newElementId
+    }
+    this.attach(insertAtStart)
+  }
+  abstract configure(): void
+  abstract renderContent(): void
+
+  private attach(insertAtBiginning: boolean) {
+    this.hostElement.insertAdjacentElement(insertAtBiginning ? 'afterbegin' : 'beforeend', this.element)
+  }
+}
+
