+++ src/components/list.ts
@@ -1,0 +1,73 @@
+import { Component } from './base.js';
+import { DragTarget } from '../models/drag-drop.js'
+import {Project, ProjectStatus} from '../models/project.js'
+import { autobind } from '../decorators/autobind.js'
+import { projectState } from '../state/project-state.js'
+import { ProjectItem}from './item.js'
+
+//Projectlist class
+export class ProjectList extends Component<HTMLDivElement,HTMLElement> implements DragTarget{
+  assignedProject: Project[]
+  constructor(private type: 'active' | 'finished') {
+    super("project-list","app",false,`${type}-projects`)
+    this.assignedProject = []
+
+    this.configure()
+    this.renderContent()
+  }
+
+  @autobind
+  dragOverHandler(event: DragEvent) { 
+    if (event.dataTransfer && event.dataTransfer.types[0] === 'text/plain') {
+      event.preventDefault()
+      const listEl = this.element.querySelector('ul')!
+      listEl.classList.add('droppable')
+    }
+    
+  }
+
+  @autobind
+  dropHandler(event: DragEvent) { 
+    const prjId = event.dataTransfer!.getData('text/plain');
+    projectState.moveProject(prjId,this.type === 'active' ? ProjectStatus.Active : ProjectStatus.Finished)
+    
+  }
+
+  @autobind
+  dragLeaveHandler(_event: DragEvent) {
+    const listEl = this.element.querySelector('ul')!
+    listEl.classList.remove('droppable')
+  }
+
+  configure() {
+    this.element.addEventListener('dragover', this.dragOverHandler)
+    this.element.addEventListener('drop', this.dropHandler)
+    this.element.addEventListener('dragleave',this.dragLeaveHandler)
+
+    projectState.addListener((projects: Project[]) => {
+      const relevantProjects = projects.filter(prj => {
+        if (this.type === 'active') {
+          return prj.status === ProjectStatus.Active
+        }
+        return prj.status === ProjectStatus.Finished
+      })
+      this.assignedProject = relevantProjects
+      this.renderProjects()
+    })
+  }
+
+  renderContent() {
+    const listId = `${this.type}-projects-list`
+    this.element.querySelector('ul')!.id = listId
+    this.element.querySelector('h2')!.textContent = this.type === 'active' ? '実行中プロジェクト' : '完了したプロジェクト'
+  }
+
+  private renderProjects() {
+    const listEl = document.getElementById(`${this.type}-projects-list`)! as HTMLUListElement
+    listEl.innerHTML = '';
+    for (const prjItem of this.assignedProject) {
+      new ProjectItem(listEl.id, prjItem)
+    }
+  }
+}
+
