+++ src/models/project.ts
@@ -1,0 +1,12 @@
+
+//project type
+export enum ProjectStatus {
+  Active,Finished
+}
+
+export class Project {
+  constructor(public id: string, public title: string, public description: string, public manday: number, public status: ProjectStatus) {
+
+  }
+}
+
