+++ src/util/validate.ts
@@ -1,0 +1,31 @@
+
+//validation
+export interface Validatable {
+  value: string | number
+  required?: boolean
+  minLength?: number
+  maxLength?: number
+  min?: number
+  max?: number
+}
+
+export function validate(validatableInput: Validatable) {
+  let isValid = true
+  if (validatableInput.required) {
+    isValid = isValid && validatableInput.value.toString().trim().length !== 0
+  }
+  if (validatableInput.minLength != null && typeof validatableInput.value === 'string') {
+    isValid && validatableInput.value.length >= validatableInput.minLength
+  }
+  if (validatableInput.maxLength != null && typeof validatableInput.value === 'string') {
+    isValid && validatableInput.value.length <= validatableInput.maxLength
+  }
+  if (validatableInput.min != null && typeof validatableInput.value === 'number') {
+    isValid && validatableInput.value >= validatableInput.min
+  }
+  if (validatableInput.max != null && typeof validatableInput.value === 'number') {
+    isValid && validatableInput.value <= validatableInput.max
+  }
+  return isValid
+}
+
